# xt_server_openApi_autogenerated.json
{
  "swagger": "2.0",
  "info": {
    "title": "XTVision server",
    "description": "All XTVsion server apis",
    "version": "1.0.0"
  },
  "host": "localhost:5001",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {},
  "paths": {
...
    "/users/": {
      "get": {
        "tags": [
          "USERS"
        ],
        "summary": "get all users",
        "description": "",
        "responses": {
          "200": {
            "description": "Returned users",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number",
                    "example": 1
                  },
                  "name": {
                    "type": "string",
                    "example": "xtvision"
                  },
                  "description": {
                    "type": "string",
                    "example": "Default user"
                  },
                  "language": {
                    "type": "string",
                    "example": "fr"
                  },
                  "time_zone": {
                    "type": "string",
                    "example": "UTC+1"
                  },
                  "createdAt": {
                    "type": "string",
                    "example": "2022-08-05T10:04:24.651Z"
                  },
                  "updatedAt": {
                    "type": "string",
                    "example": "2022-08-05T10:04:24.651Z"
                  },
                  "Roles": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string",
                          "example": "Exploitation"
                        }
                      },
                      "required": [
                        "name"
                      ]
                    }
                  },
                  "ManagementAreas": {
                    "type": "array",
                    "example": [],
                    "items": {}
                  }
                },
                "required": [
                  "id",
                  "name",
                  "language",
                  "time_zone",
                  "createdAt",
                  "updatedAt"
                ]
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "USERS"
        ],
        "summary": "Create a user",
        "description": "",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "user to create",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "string"
                },
                "password": {
                  "type": "string",
                  "example": "string"
                },
                "description": {
                  "type": "string",
                  "example": "string"
                },
                "language": {
                  "type": "string",
                  "example": "string"
                },
                "time_zone": {
                  "type": "string",
                  "example": "string"
                }
              },
              "required": [
                "name",
                "language",
                "time_zone"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returned user",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                },
                "name": {
                  "type": "string",
                  "example": "xtvision"
                },
                "description": {
                  "type": "string",
                  "example": "Default user"
                },
                "language": {
                  "type": "string",
                  "example": "fr"
                },
                "time_zone": {
                  "type": "string",
                  "example": "UTC+1"
                },
                "roleIds": {
                  "type": "array",
                  "example": [
                    1,
                    2
                  ],
                  "items": {
                    "type": "number"
                  }
                },
                "dashboardIds": {
                  "type": "array",
                  "example": [
                    1,
                    2
                  ],
                  "items": {
                    "type": "number"
                  }
                },
                "managementAreaIds": {
                  "type": "array",
                  "example": [
                    1,
                    2
                  ],
                  "items": {
                    "type": "number"
                  }
                }
              },
              "required": [
                "id",
                "name",
                "language",
                "time_zone"
              ],
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "USERS"
        ],
        "summary": "get specific user by id",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "user ID."
          }
        ],
        "responses": {
          "200": {
            "description": "Returned user by id",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                },
                "name": {
                  "type": "string",
                  "example": "xtvision"
                },
                "description": {
                  "type": "string",
                  "example": "Default user"
                },
                "language": {
                  "type": "string",
                  "example": "fr"
                },
                "time_zone": {
                  "type": "string",
                  "example": "UTC+1"
                },
                "createdAt": {
                  "type": "string",
                  "example": "2022-08-05T10:04:24.651Z"
                },
                "updatedAt": {
                  "type": "string",
                  "example": "2022-08-05T10:04:24.651Z"
                },
                "Roles": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string",
                        "example": "Exploitation"
                      }
                    },
                    "required": [
                      "name"
                    ]
                  }
                },
                "ManagementAreas": {
                  "type": "array",
                  "example": [],
                  "items": {}
                }
              },
              "required": [
                "id",
                "name",
                "language",
                "time_zone",
                "createdAt",
                "updatedAt"
              ],
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "USERS"
        ],
        "summary": "Update a user",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "user ID."
          },
          {
            "name": "user",
            "in": "body",
            "description": "user to update",
            "schema": {
              "type": "object",
              "properties": {
                "password": {
                  "type": "string",
                  "example": "string"
                },
                "description": {
                  "type": "string",
                  "example": "string"
                },
                "language": {
                  "type": "string",
                  "example": "string"
                },
                "time_zone": {
                  "type": "string",
                  "example": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "delete": {
        "tags": [
          "USERS"
        ],
        "summary": "delete one user by id",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "user ID."
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/users/filter": {
      "post": {
        "tags": [
          "USERS"
        ],
        "summary": "Get users by filter",
        "description": "",
        "parameters": [
          {
            "name": "filters",
            "in": "body",
            "description": "filter users by",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "string"
                },
                "roles": {
                  "type": "array",
                  "example": [
                    "string"
                  ],
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returned users",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number",
                    "example": 1
                  },
                  "name": {
                    "type": "string",
                    "example": "xtvision"
                  },
                  "description": {
                    "type": "string",
                    "example": "Default user"
                  },
                  "language": {
                    "type": "string",
                    "example": "fr"
                  },
                  "time_zone": {
                    "type": "string",
                    "example": "UTC+1"
                  },
                  "roleIds": {
                    "type": "array",
                    "example": [
                      1,
                      2
                    ],
                    "items": {
                      "type": "number"
                    }
                  },
                  "dashboardIds": {
                    "type": "array",
                    "example": [
                      1,
                      2
                    ],
                    "items": {
                      "type": "number"
                    }
                  },
                  "managementAreaIds": {
                    "type": "array",
                    "example": [
                      1,
                      2
                    ],
                    "items": {
                      "type": "number"
                    }
                  }
                },
                "required": [
                  "id",
                  "name",
                  "language",
                  "time_zone"
                ]
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
    "/equipments/{id}/actions/{actionType}": {
      "post": {
        "summary": "Operate an equipment to do an action in a system",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "equipment ID."
          },
          {
            "name": "actionType",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "action type."
          },
          {
            "name": "params",
            "in": "body",
            "description": "map of params to do an action",
            "schema": {
              "type": "object",
              "properties": {
                "param": {
                  "type": "string",
                  "example": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
    "/actionUsers/": {
      "post": {
        "tags": [
          "ACTION USERS"
        ],
        "summary": "create news action user",
        "description": "",
        "parameters": [
          {
            "name": "actionUser",
            "in": "body",
            "description": "new action user object",
            "schema": {
              "type": "object",
              "properties": {
                "action": {
                  "type": "string",
                  "example": "string"
                },
                "timestamp": {
                  "type": "string",
                  "example": "string"
                },
                "category": {
                  "type": "string",
                  "example": "string"
                },
                "settings": {
                  "type": "array",
                  "example": [
                    "string"
                  ],
                  "items": {
                    "type": "string"
                  }
                }
              },
              "required": [
                "action",
                "timestamp",
                "category",
                "settings"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "the id of the created user",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/actionUsers/filter": {
      "post": {
        "tags": [
          "ACTION USERS"
        ],
        "summary": "get action user by filter ",
        "description": "",
        "parameters": [
          {
            "name": "filters",
            "in": "body",
            "description": "filter action user by keyword and period",
            "schema": {
              "type": "object",
              "properties": {
                "period": {
                  "type": "array",
                  "example": [
                    "string"
                  ],
                  "items": {
                    "type": "string"
                  }
                },
                "keyword": {
                  "type": "string",
                  "example": "string"
                }
              },
              "required": [
                "period",
                "keyword"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "the id of the created user",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                },
                "action": {
                  "type": "string",
                  "example": "string"
                },
                "timestamp": {
                  "type": "string",
                  "example": "string"
                },
                "category": {
                  "type": "string",
                  "example": "string"
                },
                "settings": {
                  "type": "array",
                  "example": [
                    "settings"
                  ],
                  "items": {
                    "type": "string"
                  }
                }
              },
              "required": [
                "id",
                "action",
                "timestamp",
                "category",
                "settings"
              ],
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
  }
}